refactor: implement multi-group CPU detection and affinity handling
