+++ src/system/cpu/group_affinity/group_affinity.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone)]
+pub struct GroupAffinity {
+    pub mask: usize,
+    pub group: u16,
+}
