+++ src/system/cpu/mod.rs
@@ -1,2 +1,9 @@
 mod backend;
+mod core;
 pub mod cpu;
+pub mod group_affinity;
+mod thread;
+mod topology;
+mod vendor;
+
+// pub use cpu::;
