+++ src/system/mod.rs
@@ -1,4 +1,3 @@
-mod affinity;
 mod cpu;
 mod ioctl;
 mod kernal_driver;
